Add fzf keybindings for fish
The native installation of fzf installs fzf_key_bindings
as a symlink rather than full text. I'm including the
full text here for portability, but it should be kept
in mind that the contents of this may need to change
with future versions of fzf and fish.

It was symlinked to:

    /usr/local/Cellar/fzf/0.17.3/shell/key-bindings.fish