Add ggb for tracking history across files
Sometimes it is necessary to find a particular string
across the codebase and see when it was introduced.
For example, something like `eslint-disable` across a
project to see if it was added during a massive React
upgrade, or if it was already there in the original
codebase.

ggb is git grep + git blame, and allows for searching
for a string with it's last commit information in the
entire project.